Added helper.go and update delete.go
