+++ handler/delete.go
@@ -1,16 +1,11 @@
 package handler
 
 import (
-	"encoding/json"
 	"io"
 	"net/http"
 )
 
 type errDelete struct {
-	ErrMsg string
-}
-
-type msgErr struct {
 	ErrMsg string
 }
 
@@ -28,12 +23,4 @@ func DeleteHandler(w http.ResponseWriter, r *http.Request) {
 		io.WriteString(w, errMsg("Supports only DELETE method. Please use DELETE method."))
 		return
 	}
-}
-func errMsg(msg string) string {
-	Msg := msgErr{msg}
-	outErr, err := json.Marshal(&Msg)
-	if err != nil {
-		return `{"Error code": "-1"}`
-	}
-	return string(outErr)
 }
